Rename StepVelocity to Velocity
Move it its own file
Update examples
