+++ Example Mac/ViewController.swift
@@ -39,7 +39,7 @@ class ViewController: NSViewController {
         noteValue: .quarter),
       bpm: 80)
 
-    let bassVolume = MIDISequencerStepVelocity.standard(100)
+    let bassVolume = MIDISequencerVelocity.standard(100)
     let bass = MIDISequencerTrack(
       name: "Bass",
       midiChannels: [1],
@@ -86,7 +86,7 @@ class ViewController: NSViewController {
           velocity: bassVolume),
       ])
 
-    let chordsVolume = MIDISequencerStepVelocity.standard(100)
+    let chordsVolume = MIDISequencerVelocity.standard(100)
     let chords = MIDISequencerTrack(
       name: "Chords",
       midiChannels: [2],
